# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ReZ Commerce Media Platform API
  description: |
    ## Overview

    The ReZ Commerce Media Platform provides a unified API for managing advertising campaigns,
    dynamic pricing, wallet transactions, and analytics across multiple channels including
    In-App, DOOH, QR, Broadcast, and Offline advertising.

    ## Key Features

    - **Dynamic Pricing Engine**: AI-powered pricing with real-time demand, competition, and
      seasonal multipliers
    - **Unified Campaign Management**: Create and manage campaigns across multiple ad channels
    - **Merchant Wallet**: Integrated wallet system for campaign budgeting and payments
    - **Real-time Analytics**: Track impressions, clicks, conversions, and ROI
    - **Webhook Events**: Receive real-time notifications for campaign and payment events

    ## Architecture

    ```
    Client -> API Gateway -> Services
                          ├── Pricing Engine (Port 4008)
                          ├── Ads Service (Port 4007)
                          ├── Ad Campaigns (Port 4008)
                          └── Wallet Service
    ```

  version: 1.0.0
  contact:
    name: ReZ Developer Support
    email: api-support@rez.money
    url: https://developers.rez.money
  license:
    name: Proprietary
    url: https://rez.money/terms
servers:
  - url: https://api.rez.money/v1
    description: Production server
  - url: https://api-staging.rez.money/v1
    description: Staging server
...
paths:
...
  /campaigns:
    get:
      tags:
        - Campaigns
      summary: List campaigns
      description: |
        Retrieve a paginated list of campaigns for the authenticated merchant.
        Supports filtering by status, type, and date range.
      operationId: listCampaigns
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          description: Filter by campaign status
          schema:
            type: string
            enum: [active, paused, completed, cancelled, draft]
        - name: type
          in: query
          description: Filter by ad type
          schema:
            type: string
            enum: [banner, feed, search, store, push, whatsapp, email, dooh, offline, qr]
        - name: startDate
          in: query
          description: Filter campaigns starting from this date
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter campaigns ending before this date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Campaigns retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Campaigns
      summary: Create campaign
      description: |
        Create a new advertising campaign. The campaign will be validated
        against minimum spend requirements and wallet balance.

        **Campaign Lifecycle:**
        1. Draft -> Active (after payment confirmation)
        2. Active -> Paused (manual or automatic)
        3. Active -> Completed (end date reached or budget exhausted)
        4. Any state -> Cancelled (manual cancellation)
      operationId: createCampaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '400':
          description: Invalid request or insufficient wallet balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '402':
          description: Payment required - insufficient wallet balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /campaigns/{id}:
    get:
      tags:
        - Campaigns
      summary: Get campaign details
      description: Retrieve detailed information about a specific campaign.
      operationId: getCampaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
      responses:
        '200':
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Campaigns
      summary: Update campaign
      description: |
        Update campaign settings. Only certain fields can be modified
        depending on campaign status.
      operationId: updateCampaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCampaignRequest'
      responses:
        '200':
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '400':
          description: Invalid update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Campaigns
      summary: Delete campaign
      description: Delete a campaign. Only draft campaigns can be deleted.
      operationId: deleteCampaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
      responses:
        '204':
          description: Campaign deleted
        '400':
          description: Cannot delete non-draft campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from the `/auth/login` endpoint.
        Include as: `Authorization: Bearer <token>`
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code (e.g., "INVALID_REQUEST", "UNAUTHORIZED")
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        requestId:
          type: string
          description: Request ID for support reference
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages

    # --------------------------------------------------------------------------
    # Authentication
    # --------------------------------------------------------------------------
...
    Location:
      type: object
      properties:
        city:
          type: string
          description: City name
        area:
          type: string
          description: Specific area or neighborhood
        tier:
          type: string
          enum: [tier1, tier2, tier3]
          description: City tier classification
        coordinates:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
    Audience:
      type: object
      properties:
        segment:
          type: string
          description: Audience segment (e.g., "young_professionals")
        income:
          type: string
          enum: [low, medium, high]
          description: Income level
        ageGroup:
          type: string
          description: Age group target
        category:
          type: string
          enum: [luxury, real_estate, restaurant, events, healthcare, retail, services]
...
    CreateCampaignRequest:
      type: object
      required:
        - name
        - adType
        - budget
        - goalType
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Campaign name
        adType:
          type: string
          enum: [banner, feed, search, store, push, whatsapp, email, dooh, offline, qr]
        placement:
          type: string
          description: Specific placement ID
        budget:
          type: number
          minimum: 0
          description: Campaign budget in INR
        duration:
          type: integer
          description: Campaign duration in days
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        location:
          $ref: '#/components/schemas/Location'
        targeting:
          $ref: '#/components/schemas/Audience'
        creative:
          $ref: '#/components/schemas/CampaignCreative'
        goalType:
          type: string
          enum: [awareness, clicks, conversions, sales, footfall, qr_scans, leads]
        campaignMode:
          type: string
          enum: [auction, reserved]
          default: auction
    CampaignCreative:
      type: object
      properties:
        headline:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          format: uri
        ctaText:
          type: string
        landingUrl:
          type: string
          format: uri
    UpdateCampaignRequest:
      type: object
      properties:
        name:
          type: string
        budget:
          type: number
        status:
          type: string
          enum: [paused, active]
        endDate:
          type: string
          format: date
        targeting:
          $ref: '#/components/schemas/Audience'
        creative:
          $ref: '#/components/schemas/CampaignCreative'
    CampaignResponse:
      type: object
      properties:
        id:
          type: string
        merchantId:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed, cancelled]
        adType:
          type: string
        budget:
          type: number
        spent:
          type: number
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CampaignListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CampaignResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
security:
  - bearerAuth: []